Refactor: move non-endpoint code to common directory
