+++ src/common/repositories/IEntity.ts
