+++ src/common/repositories/fav.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { Repository } from './repository';
-import { Favorites } from '../favs/entities/fav.entity';
+import { Favorites } from '../../favs/entities/fav.entity';
 
 const favouritesRepository: Array<Favorites> = [new Favorites()];
